+++ pkg/ring/kv/memberlist/memberlist_client.go
@@ -72,7 +72,12 @@ func (c *Client) Get(ctx context.Context, key string) (any, error) {
 
 // Delete is part of kv.Client interface.
 func (c *Client) Delete(ctx context.Context, key string) error {
-	return errors.New("memberlist does not support Delete")
+	err := c.awaitKVRunningOrStopping(ctx)
+	if err != nil {
+		return err
+	}
+
+	return c.kv.Delete(key)
 }
 
 // CAS is part of kv.Client interface
@@ -677,6 +682,15 @@ func (m *KV) get(key string, codec codec.Codec) (out any, version uint, err erro
 	}
 
 	return v.value, v.version, nil
+}
+
+func (m *KV) Delete(key string) error {
+	// TODO(Sungjin1212): Mark as delete and broadcast to peers
+	m.storeMu.Lock()
+	defer m.storeMu.Unlock()
+
+	delete(m.store, key)
+	return nil
 }
 
 // WatchKey watches for value changes for given key. When value changes, 'f' function is called with the